+++ scripts/addie
@@ -51,8 +51,8 @@ from addie.utilities.ipts_file_transfer_dialog import IptsFileTransferDialog
 
 from addie.initialization.init_step1 import InitStep1
 
-from addie.step1_handler.step1_gui_handler import Step1GuiHandler
-from addie.step1_handler.run_step1 import RunStep1
+from addie.autoNOM.step1_gui_handler import Step1GuiHandler
+from addie.autoNOM.run_step1 import RunStep1
 from addie.step2_handler.load_table_intermediate_step_interface import loadTableIntermediateStepInterface
 from addie.step2_handler.undo_handler import UndoHandler
